feat: 댓글 등록 API 개발 (#34)

+++ src/main/java/kr/adapterz/community/domain/member/repository/MemberRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import kr.adapterz.community.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
     Optional<Member> findByEmail(String email);
     boolean existsByEmail(String email);
     boolean existsByNickname(String nickname);
